Getting closer
Need to figure out how to exclude Polymer app from build
