+++ rollup.config.js
@@ -1,21 +1,11 @@
 import resolve from 'rollup-plugin-node-resolve';
 import typescript from 'rollup-plugin-typescript2';
 
-export default [
-  {
-    input: 'src/editor.ts',
-    output: {
-      file: 'boilerplate-card-editor.js',
-      format: 'es',
-    },
-    plugins: [resolve(), typescript()],
+export default {
+  input: ['src/boilerplate-card-editor.ts', 'src/boilerplate-card.ts'],
+  output: {
+    dir: './dist',
+    format: 'es',
   },
-  {
-    input: 'src/index.ts',
-    output: {
-      file: 'boilerplate-card.js',
-      format: 'es',
-    },
-    plugins: [resolve(), typescript()],
-  },
-];
+  plugins: [resolve(), typescript()],
+};
