+++ src/boilerplate-types.ts
@@ -1,0 +1,7 @@
+// TODO Add your configuration elements here for type-checking
+export interface BoilerplateConfig {
+  type: string;
+  name?: string;
+  show_warning?: boolean;
+  show_error?: boolean;
+}
